+++ encoding/toml/parser_test.ts
@@ -1,11 +1,12 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import { test } from "../testing/mod.ts";
-import { assertEquals } from "../testing/asserts.ts";
-import { existsSync } from "../fs/exists.ts";
-import { readFileStrSync } from "../fs/read_file_str.ts";
+import { test } from "../../testing/mod.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { existsSync } from "../../fs/exists.ts";
+import { readFileStrSync } from "../../fs/read_file_str.ts";
 import { parse, stringify } from "./parser.ts";
-import * as path from "../fs/path/mod.ts";
-const testFilesDir = path.resolve("toml", "testdata");
+import * as path from "../../fs/path/mod.ts";
+
+const testFilesDir = path.resolve("encoding", "toml", "testdata");
 
 function parseFile(filePath: string): object {
   if (!existsSync(filePath)) {
